requirements.txt
gradio==3.50.2
openai>=1.99.1
Pillow>=10.0.0
python-docx>=1.1.0
numpy>=1.24.0